requirements.txt
# Tested with Python 3.8 ----------------------------
# Tested with macOS 11.6.2 --------------------------

addict==2.4.0
elasticdeform==0.4.9
Flask==2.1.2
flask_sqlalchemy==2.5.1
imageio==2.19.1
matplotlib==3.5.2
numpy==1.21.6
PyYAML==6.0
scipy==1.7.3
scikit-image==0.19.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
